Remove '(if any)' from username requirement.
It's a seriously time consuming job to reverse engineer someones account
from their scrawled handwriting, and impossible in the case of name
collisions.

This might lead to some parents / other attendees asking whether they
need accounts, but replying to those emails is going to be less
irritating than dealing with lots of pieces of paper.

Fixes #1121